Update to 5.10.13.1
Move to the new CBL-Mariner kernel source location and use the latest
5.10.13.1 version.

As part of the upgrade to 5.10.13.1, we can remove some out-of-tree
patches since these patches have been merged into upstream.

Additionally, we need to remove aarch64 module.lds because this file
is no longer checked in as part of the source tree. See this upstream
commit for more details:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=596b0474d3d9b1242eab713f84d8873f9887d980

Signed-off-by: Chris Co <chrco@microsoft.com>